changes to the UI
