+++ src/index.css
@@ -9,6 +9,7 @@
 
 body {
   font-family: 'Didact Gothic', sans-serif;
+  background: var(--body-bg);
 }
 
 .boxes {
@@ -17,12 +18,15 @@ body {
   width: 100%;
   gap: 10px;
   background-color: #fff;
+  background: var(--body-bg);
 }
 
 :root {
   --color-text: #fff;
-  --color-front: hsl(250deg, 100%, 40%);
-  --color-back: hsl(252deg, 100%, 60%);
+  --color-back: #1e212c;
+  --color-front: #050b18;
+  --color-btn: #644dda;
+  --body-bg: #2a2c44;
 }
 
 .series {
@@ -56,7 +60,7 @@ body {
 .series:hover::after {
   transition: transform 500ms;
   background: var(--color-back);
-  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.1);
+  box-shadow: 3px 3px 3px rgba(0, 0, 0, 0.1);
   border-radius: 0 0 6px 0;
   width: 80px;
   height: 80px;
@@ -136,10 +140,10 @@ body {
 }
 
 footer {
-  position: absolute;
-  left: 0.5rem;
-  bottom: 0.5rem;
-  right: 0.5rem;
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  right: 0;
   border: 3px solid black;
   font-size: 1.2rem;
   box-shadow: 0 0 10px darkgrey;
@@ -185,4 +189,13 @@ footer p {
 .line2 {
   width: 12px;
   height: 3px;
+}
+
+button{
+  background-color: var(--color-btn);
+  padding: 12px;
+  color: var(--color-text);
+  border-radius: 5px;
+  outline: 0;
+  border: none;
 }
