+++ src/index.js
@@ -1,8 +1,4 @@
 import './index.css';
 import getSeries from './module/getSeries.js';
-import displaySeries from './module/displaySeries.js';
 
-window.onload = async () => {
-  const items = await getSeries();
-  displaySeries(items);
-};
+window.addEventListener("load" , getSeries)
